Little reminder about follow requests
Our friend.camp users found it weird that on mobile, follow requests are
buried. As admin, I didn't have the skill to add follow requests to the
notifications tab, but this commit makes it so a reminder of unread
follow requests will appear beneath the TOOT button if the user has
their account set to locked.