Add a "never ask me again" to Clear Notifications modal
This adds a "Never ask me again" checkbox to the notification
confirmation clear modal. When the user checks the box (regardless of
whether they confirm the actual clearing), a variable is stored in
LocalStorage that remembers it was checked (or subsequently unchecked).
If the box is checked and the variable in LocalStorage is `true`, then
the modal won't appear the next time the user hits "Clear
Notifications".

This means that it's possible to get in a weird place where you select
the checkbox, cancel the modal, but still never see the modal again. I'm
okay with this. You may not be.

Requested by @li@friend.camp