[FC] Images go behind CW cut, not toggled off
This commit changes the way media is rendered in the timeline.
Previously a status with media was formatted like this:

StatusContainer
  StatusContent
  MediaGallery

Now it's formatted like

StatusContainer
  StatusContent
    MediaGallery

So when you toggle statuscontent, you also toggle the MediaGallery. This
required moving the MediaGallery stuff into the status_content.js
from status.js. It also required a change to how detailed statuses work.
As of this commit, detailed statuses still require the extra click
because it's complicated and this solves most of my users' problems with
this whole thing.

CWs with an image show the "eye" icon next to them so you know an image
is coming.